projects/05-event-stream-analytics/requirements.txt
pandas>=2.2
pyarrow>=16.0
duckdb>=1.1
requests>=2.31
python-dotenv>=1.0
kafka-python>=2.0
pytest>=8.0
ipykernel>=6.29
